Override CPU info for ARM variants
The current docker and balenaEngine binaries refer
to /proc/cpuinfo to determine the ARM variant on startup.

When using QEMU to emulate ARM on x86, this file reflects the host
CPU and not the emulated CPU.

Similarly, when running arm7l runners on arm64 hosts we want to force
the Docker to detect the target ARM variant from the container.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>