Add e2e workflow that runs on draft runners
The docker compose tests in Flowzone will bring
the runners online with a dynamic tag specific
to the commit hash.

The e2e workflow will wait and execute on the
draft runners when they come online.

When the e2e job is complete, the runner
will automatically teardown, completing the
docker compose test step.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>