Refactor secret and variable management to common functions
The goal is to categorize all env vars as either secrets or variables
before loading them in each service where they are required.

This allows better control of secrets, like when they are loaded
and when they are deleted from the filesystem, in a common way that can
be shared with the VM runner environment.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>