Use a nodejs script to configure dockerd and remove remco
Golang and pongo templates are simply not the best
way to generate complex JSON from templates. Also the documentaion
for both is lacking and a simple node script can do the job.

If we find a need for templated config files for
other services we can introduce a confd s6 service.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>