Use chroot and overlayfs for container runtime filesystem
The main goal of this init script is to set up the filesystem
in a way that changes are not persisted across container restarts.

Using overlayfs allows faster startup as we don't have to copy
the root filesystem on every boot, but we can still isolate
any runtime filesystem changes in a separate layer.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>