Use pivot_root and read_only for ephemeral rootfs
The previous method was resulting in corruption
of the underlying overlayfs and resulted in unwritable
directories like /usr

Instead create a new rootfs in a volume and use pivot_root
to switch to it on container start.

This allows the original container root to be set to read-only.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>