Fix update-lock dependency ordering
The update-locks service has no depedencies because
it's just a shell process regularly checking to see if the
worker process exists.

Instead, the runner should only start if the update-locks
loop is already started.

Also, having both services be a dependency of the entrypoint
meant that removing the runner service didn't actually stop
it from running as they depended on each other.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>